Moved the mapBlocksToRenderComponents logic to the PostContent component
