+++ lib/blocks.ts
@@ -1,5 +1,5 @@
-import { createElement, ReactNode } from 'react';
-import { ContentBlock, ContentBlockAttribute } from '@/graphql/generated';
+import { ReactNode } from 'react';
+import { ContentBlockAttribute } from '@/graphql/generated';
 import {
 	ClassicEditorBlock,
 	Heading,
@@ -52,42 +52,7 @@ function mapBlockNamesToComponents ( override?: BlocksToComponentsProps ): Recor
 	}
 }
 
-type BlocksToRenderComponentsProps = {
-	blocks: ContentBlock[],
-	blockMapOverrides?: Record<string, ReactNode>,
-};
-
-function mapBlocksToRenderComponents ( { blocks, blockMapOverrides = {  } } : BlocksToRenderComponentsProps ) {
-	return (
-		blocks.map( ( block, i ) => {
-			const attributesProps = mapAttributesToProps( block.attributes || [] );
-			const defaultProps = { key: `block-${i}`, block };
-
-			const components = mapBlockNamesToComponents( blockMapOverrides );
-
-			if ( Object.keys( components ).includes( block.name ) ) {
-				return createElement(
-					components[block.name] as string,
-					{ ...defaultProps, ...attributesProps },
-				);
-
-			// In development, highlight unsupported blocks so that they get
-			// visibility with developers.
-			} else if ( 'development' === process.env.NODE_ENV ) {
-				return createElement(
-					components['unsupported'] as string,
-					defaultProps,
-				);
-			}
-
-			// In production, ignore unsupported blocks.
-			return null;
-		})
-	)
-}
-
 export {
 	mapAttributesToProps,
 	mapBlockNamesToComponents,
-	mapBlocksToRenderComponents,
 };
